Squash partial rewrite branch and move into an orphan branch Squashed and orphaned a local branch to avoid problems with incorrect license headers, deleted dependencies, etc...
This is a partial rewrite/refactored version of OpenOlympus 0.0.1.
The web server is now used to provide a JSON API to a frontend written
in Angular and Bootstrap. This should make it easier to test. The
backend now utilises @JsonView and Spring Security expressions. JOOQ
is now used to generate native queries.